build(win32): stop three shadowing header sets from hiding the Windows SDK
Follow-up to cc6ba925d. That commit found CompatLib/Include's shim windows.h
shadowing the real Windows SDK and cut the win32/x86 build from 5529 errors to
1490. The remaining errors were dominated by the same class of bug from two
more directories, plus a handful of genuinely POSIX-only sources.

cmake/dx8.cmake - SAGE_USE_DX8 had no default, so Windows built as Linux
  Nothing in the tree ever declared SAGE_USE_DX8: no option(), no set(). The
  only writers are the linux64-deploy, macos-vulkan and ios-vulkan presets, and
  all three set it OFF. Every Windows preset left it undefined, so `if(SAGE_USE_DX8)`
  was false and Windows fell through to the final else() - the LINUX arm - which
  FetchContent-downloads dxvk-native-2.6-steamrt-sniper.tar.gz. That is DXVK's
  Linux package: include/dxvk ships DXVK's own stand-in for the Win32 API,
  including windows.h (just windows_base.h + unknwn.h, no _WIN32 guard) and
  windows_base.h. CompatLib hangs that directory off the d3d8lib INTERFACE, so
  it reached the -I line of all 71 failing cl.exe invocations ahead of the SDK.
  Consequences: minwindef.h never ran, so FAR and PASCAL were undefined and the
  SDK's own winsock.h could not parse (522 errors, C2146 "missing ';' before
  identifier 'FAR'" at winsock.h:74 and C2086 "'char hostent::FAR': redefinition"
  at :165); windows_base.h's `typedef uint32_t ULONG/DWORD` collided with
  WWLib/bittype.h's `unsigned long` (88x C2371); and CRITICAL_SECTION, LPCTSTR,
  GetTickCount, RegOpenKeyEx, EXCEPTION_POINTERS and friends all vanished.
  Windows now defaults SAGE_USE_DX8=ON and takes the min-dx8-sdk arm that was
  written for it and never selected. min-dx8-sdk ships d3d8/d3dx8/dinput headers
  and MSVC .libs and crucially no windows.h. DXVK on Windows is not a header
  package at all - it is a drop-in d3d8.dll, and dx8wrapper.cpp already resolves
  Direct3DCreate8 via LoadLibrary + GetProcAddress. Forcing SAGE_USE_DX8=OFF on
  Windows is now a FATAL_ERROR instead of a silent fallthrough into the Linux arm.
  Also pins d3dx8.lib by absolute path under MSVC: min-dx8-sdk asks for the bare
  name "d3dx8", but CompatLib creates an empty INTERFACE target with that exact
  name on 32-bit Windows, and a CMake target always beats a library file name -
  so the .lib silently dropped off the link line.

WWVegas/CMakeLists.txt + WWAudio/ - a stub mmsystem.h on 2190 include paths
  WWAudio/mmsystem.h was a Linux/OpenAL stub whose entire body sits inside
  `#if defined(SAGE_USE_OPENAL) && !defined(_WIN32)` - an empty file on Windows.
  WWAudio is listed as an include dir on core_wwcommon, which puts it on the -I
  line of 2190 of 2718 translation units, and on MSVC a /I directory out-ranks
  the SDK include path. Every `#include <mmsystem.h>` in the Windows build
  therefore resolved to that empty file, leaving Miles' mss.h:205 without
  LPHWAVEOUT/LPWAVEFORMAT (38x C2061) and cascading through WWAudio.h,
  AudibleSound.h and Utils.h. Moved to WWAudio/compat/ (body unchanged) and
  exported that directory only when NOT WIN32, so Windows resolves the genuine
  SDK header while macOS/iOS/Linux keep the stub. Miles itself needed no change:
  the stub compiles clean against the real SDK on win32/x86 once the shadow is
  gone. Same containment pattern as the CompatLib/Include scoping in cc6ba925d.

WWDownload/ftp.h, FTP.cpp - Windows arm was missing vocabulary the POSIX arm had
  These sources were only ever built on POSIX, where windows_compat.h supplies
  strlcpy/strlcat, ARRAY_SIZE, socklen_t and <time.h> for free. The `#ifdef _WIN32`
  arm supplied none of them, so the code referenced identifiers that exist only
  off-Windows. Mirrors the same names from the engine's own headers, and adds
  <sys/stat.h> for the _S_IREAD/_S_IWRITE used by _chmod in Prepare_Directories().
  These failures were invisible before because every affected TU aborted at the
  100-error limit inside winsock.h first.

GameNetwork/udp.h - socklen_t on the Winsock 1.1 arm
  udp.cpp passes socklen_t to getsockname()/getsockopt(). Winsock 1.1 has no such
  type; it only appears in <ws2tcpip.h>, which cannot be included here because it
  drags winsock2.h in on top of winsock.h. The Winsock 1.1 prototypes take int*,
  so `typedef int socklen_t` is byte-identical to what <ws2tcpip.h> declares and
  stays legal if that header ever does arrive.

WWDebug/wwprofile.cpp - thread_compat.h is the POSIX shim
  Sole engine source hitting C1083 on thread_compat.h, which lives in
  CompatLib/Include - deliberately off the Windows path since cc6ba925d. Win32
  has ::GetCurrentThreadId() natively, so the _WIN32 arm uses the stock upstream
  declaration. Same guard style as WWLib/thread.cpp and GameEngine Debug.cpp, the
  only other consumers.

WWLib/registry.{h,cpp} - HKEY truncated into an int
  RegistryClass::Key stored an HKEY - a pointer - in an int, guarded by a runtime
  `assert(sizeof(HKEY) == sizeof(int))` that compiles away under NDEBUG, i.e. in
  the release config actually being built. Harmless at 32-bit, corrupt at 64-bit:
  compiling the original at x64 emits 19 pointer-truncation warnings (C4311/C4302/
  C4312). Key is now intptr_t with a static_assert, per jmarshall's Win64 port.

macOS/iOS safety: proven, not argued. build-macos-zh.sh (preset macos-vulkan,
full configure + build) exits 0 and links GeneralsMD/GeneralsXZH. Every source
hunk sits inside a pre-existing #ifdef _WIN32 arm whose #else is byte-identical
to before; the two ungated changes (registry.h's intptr_t member and <stdint.h>)
are inert off-Windows because the non-Windows RegistryClass stub only ever sets
Key to 0. Every CMake addition is WIN32-gated or NOT WIN32-gated, and the DXVK
arms for macOS and Linux are untouched. The WWAudio/compat include dir reaches
exactly the same 2190 translation units the old WWAudio entry did - verified by
counting it in the regenerated compile_commands.json, not by inspection.

+++ Core/Libraries/Source/WWVegas/WWLib/registry.h
@@ -39,6 +39,7 @@
 #include "Vector.h"
 #include "wwstring.h"
 #include "widestring.h"
+#include <stdint.h>	// intptr_t, for the pointer-wide RegistryClass::Key
 
 class INIClass;
 
@@ -107,7 +108,10 @@ private:
 	static void Save_Registry_Values(HKEY key, char *path, INIClass *ini);
 
 
-	int	Key;
+	// GeneralsX @bugfix Claude 27/07/2026 Key holds a HKEY, which is a pointer. An int truncates it
+	// on x64 (harmless at 32-bit, corrupt everywhere else), so store it pointer-wide. Inert on the
+	// macOS/iOS arm: the non-Windows RegistryClass stub only ever initialises this to 0.
+	intptr_t	Key;
 	bool	IsValid;
 
 	//
